backend/requirements.txt
# API + Noxis brain (Noxis.py at repo root). Trim optional packages if you need a slimmer image.
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
groq>=0.4.0
openai>=1.0.0
google-generativeai>=0.4.0
requests>=2.31.0
Pillow>=10.0.0
